+++ lib/jekyll/draft.rb
@@ -18,6 +18,11 @@ module Jekyll
       File.join(source, dir, '_drafts')
     end
 
+    # The path to the draft source file, relative to the site source
+    def relative_path
+      File.join(@dir, '_drafts', @name)
+    end
+
     # Extract information from the post filename.
     #
     # name - The String filename of the post file.