+++ test/test_filters.rb
@@ -28,7 +28,7 @@ class TestFilters < Test::Unit::TestCase
     end
 
     should "markdownify with simple string" do
-      assert_equal "<p>something <strong>really</strong> simple</p>", @filter.markdownify("something **really** simple")
+      assert_equal "<p>something <strong>really</strong> simple</p>\n", @filter.markdownify("something **really** simple")
     end
 
     should "convert array to sentence string with no args" do
@@ -131,7 +131,7 @@ class TestFilters < Test::Unit::TestCase
             assert_equal 2, g["items"].size
           when ""
             assert g["items"].is_a?(Array), "The list of grouped items for '' is not an Array."
-            assert_equal 7, g["items"].size
+            assert_equal 9, g["items"].size
           end
         end
       end